backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
supabase==1.0.3
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
openai==0.28.1
pydantic==2.4.2
python-dotenv==1.0.0